# src/internal/api/server.go
package api

import (
	"AutoAnimeDownloader/src/internal/daemon"
	"AutoAnimeDownloader/src/internal/files"
	"AutoAnimeDownloader/src/internal/frontend"
	"AutoAnimeDownloader/src/internal/logger"
	"AutoAnimeDownloader/src/internal/torrents"
	"context"
	"io"
	"io/fs"
	"net/http"
	"strings"
	"sync"
	"time"

	httpSwagger "github.com/swaggo/http-swagger"
)

// FileManagerInterface defines the interface for file management operations
// This interface matches daemon.FileManagerInterface
type FileManagerInterface interface {
	LoadConfigs() (*files.Config, error)
	SaveConfigs(config *files.Config) error
	LoadSavedEpisodes() ([]files.EpisodeStruct, error)
	SaveEpisodesToFile(episodes []files.EpisodeStruct) error
	UpsertEpisodes(episodes []files.EpisodeStruct) error
	DeleteEpisodesFromFile(keys []files.EpisodeKey) error
	DeleteEmptyFolders(completedAnimeSaveFolder string) error
	LoadBlockedEpisodes() ([]files.EpisodeKey, error)
	BlockEpisode(key files.EpisodeKey) error
	UnblockEpisode(key files.EpisodeKey) error
	UnmanageEpisode(key files.EpisodeKey) error
	LoadAllAnimeSettings() (map[int]files.AnimeSettings, error)
	LoadAnimeSettings(animeID int) (*files.AnimeSettings, error)
	SaveAnimeSettings(animeID int, settings files.AnimeSettings) error
	LoadStandaloneAnimes() ([]int, error)
	AddStandaloneAnime(mediaID int) error
	RemoveStandaloneAnime(mediaID int) error
}

type Server struct {
	http.Server
	State         *daemon.State
	FileManager   FileManagerInterface
	StartLoopFunc func(daemon.StartLoopPayload) *daemon.LoopControl
	WSManager     *WebSocketManager
	JobQueue      *daemon.JobQueue
	Torrents      torrents.TorrentBackend
	Librarian     files.Librarian

	mu                 sync.Mutex
	currentLoopControl *daemon.LoopControl

	// checks tracks the manual-verification goroutines started by handleCheck. The endpoint
	// is fire-and-forget by design, so production never waits on it — the tests do
	// (waitForChecks). A verification that outlives its test keeps calling into package-level
	// globals the next test is busy replacing (anilist's httpDo swap, the logger), and it
	// keeps hitting the network and the real disk; -race then reports the collision against
	// whichever test happens to run next, far from the one that actually leaked it.
	checks sync.WaitGroup
}

func NewServer(port string, state *daemon.State, fileManager FileManagerInterface, startLoopFunc func(daemon.StartLoopPayload) *daemon.LoopControl) *Server {
	wsManager := NewWebSocketManager()

	// Set state getter for WebSocket manager
	wsManager.SetStateGetter(func() (daemon.Status, time.Time, bool) {
		return state.GetAll()
	})

	server := &Server{
		State:         state,
		FileManager:   fileManager,
		StartLoopFunc: startLoopFunc,
		WSManager:     wsManager,
	}

	server.Server = http.Server{
		Addr:    port,
		Handler: nil, // Will be configured in SetupRoutes
	}

	return server
}

func (s *Server) SetupRoutes() *http.ServeMux {
	mux := http.NewServeMux()

	// API routes - apply JSON middleware only to API routes
	apiMux := http.NewServeMux()
	apiMux.HandleFunc("/api/v1/status", handleStatus(s))
	apiMux.HandleFunc("/api/v1/config", handleConfig(s))
	apiMux.HandleFunc("/api/v1/config/priorities/defaults", handlePriorityDefaults(s))
	apiMux.HandleFunc("/api/v1/animes", handleAnimes(s))
	apiMux.HandleFunc("/api/v1/animes/{id}/episodes", handleAnimeEpisodes(s))
	apiMux.HandleFunc("/api/v1/animes/{id}/episodes/{episodeNumber}/download", handleDownloadEpisode(s))
	apiMux.HandleFunc("/api/v1/animes/{id}/episodes/{episodeNumber}/redownload", handleRedownloadEpisode(s))
	apiMux.HandleFunc("/api/v1/animes/{id}/episodes/{episodeNumber}/release", handleReleaseEpisode(s))
	apiMux.HandleFunc("/api/v1/animes/{id}/episodes/{episodeNumber}/replace", handleReplaceEpisodeWithMagnet(s))
	apiMux.HandleFunc("/api/v1/animes/{id}/episodes/{episodeNumber}", handleDeleteEpisode(s))
	apiMux.HandleFunc("/api/v1/animes/{id}/replace", handleReplaceAnimeWithMagnet(s))
	apiMux.HandleFunc("/api/v1/animes/{id}/settings", handleAnimeSettings(s))
	apiMux.HandleFunc("/api/v1/anilist/search", handleAniListSearch(s))
	apiMux.HandleFunc("/api/v1/standalone-animes", handleStandaloneAnimeAdd(s))
	apiMux.HandleFunc("/api/v1/standalone-animes/{id}", handleStandaloneAnimeRemove(s))
	apiMux.HandleFunc("/api/v1/check", handleCheck(s))
	apiMux.HandleFunc("/api/v1/daemon/start", handleDaemonStart(s))
	apiMux.HandleFunc("/api/v1/daemon/stop", handleDaemonStop(s))
	apiMux.HandleFunc("/api/v1/logs", handleLogs(s))
	apiMux.HandleFunc("/api/v1/torrents", handleTorrents(s))
	// Single pattern for every method on this path: Go 1.22+ ServeMux patterns without a
	// method prefix match all verbs, so handleTorrentDelete's own method check is what turns
	// a non-DELETE request into a 405 instead of the mux ever seeing an unmatched pattern.
	// This does not collide with "/api/v1/torrents" (different segment count), nor with the
	// "/pause", "/resume", "/announce", "/prioritize" sub-paths below (a bare "{hash}" pattern
	// only matches a single path segment), nor with the literal "/api/v1/torrents/prioritize"
	// batch route: Go 1.22+ gives a literal segment precedence over a wildcard, and no info
	// hash is the string "prioritize" anyway (they are 40 hex chars).
	apiMux.HandleFunc("/api/v1/torrents/prioritize", handleTorrentsPrioritize(s))
	apiMux.HandleFunc("/api/v1/torrents/{hash}", handleTorrentDelete(s))
	apiMux.HandleFunc("/api/v1/torrents/{hash}/pause", handleTorrentPause(s))
	apiMux.HandleFunc("/api/v1/torrents/{hash}/resume", handleTorrentResume(s))
	apiMux.HandleFunc("/api/v1/torrents/{hash}/announce", handleTorrentAnnounce(s))
	apiMux.HandleFunc("/api/v1/torrents/{hash}/prioritize", handleTorrentPrioritize(s))
	apiMux.HandleFunc("/api/v1/notifications/webhooks/{name}/test", handleNotificationWebhookTest(s))

	// WebSocket route (no JSON middleware)
	mux.HandleFunc("/api/v1/ws", s.handleWebSocket())

	// Apply middlewares to API routes
	mux.Handle("/api/", ApplyMiddlewares(apiMux))
	mux.Handle("/swagger/", ApplyMiddlewares(httpSwagger.Handler(
		httpSwagger.URL("/swagger/doc.json"),
	)))

	// Static files (frontend) - apply only CORS and logging, not JSON middleware
	staticHandler := corsMiddleware(loggingMiddleware(s.handleStaticFiles()))
	mux.Handle("/", staticHandler)

	return mux
}

func (s *Server) handleStaticFiles() http.HandlerFunc {
	// Use embedded frontend filesystem
	// The embed includes dist/*, so we need to use fs.Sub to get the dist directory
	distFS, err := fs.Sub(frontend.DistFS, "dist")
	if err != nil {
		logger.Logger.Fatal().
			Err(err).
			Msg("Failed to create sub filesystem for frontend dist")
	}

	logger.Logger.Info().
		Msg("Serving frontend from embedded filesystem")

	fileServer := http.FileServer(http.FS(distFS))

	return func(w http.ResponseWriter, r *http.Request) {
		// Don't serve static files for API routes
		if strings.HasPrefix(r.URL.Path, "/api/") || strings.HasPrefix(r.URL.Path, "/swagger/") {
			http.NotFound(w, r)
			return
		}

		path := r.URL.Path

		// For SPA routing: serve index.html for paths without file extensions
		if path == "/" || (!strings.Contains(path, ".") && path != "/") {
			// Check if the requested path exists as a file
			if path != "/" {
				file, err := distFS.Open(strings.TrimPrefix(path, "/"))
				if err == nil {
					// File exists, serve it normally
					file.Close()
					fileServer.ServeHTTP(w, r)
					return
				}
			}
			// File doesn't exist or it's root path, serve index.html for SPA routing
			// Read and serve the file directly to avoid redirect loops
			file, err := distFS.Open("index.html")
			if err != nil {
				http.NotFound(w, r)
				return
			}
			defer file.Close()

			// Set content type
			w.Header().Set("Content-Type", "text/html; charset=utf-8")

			// Copy file content to response
			_, err = io.Copy(w, file)
			if err != nil {
				logger.Logger.Error().Err(err).Msg("Failed to serve index.html")
			}
			return
		}

		// Serve the file normally for paths with extensions
		fileServer.ServeHTTP(w, r)
	}
}

func (s *Server) Start() error {
	mux := s.SetupRoutes()
	s.Handler = mux

	logger.Logger.Info().
		Str("addr", s.Addr).
		Msg("Starting API server")

	return s.ListenAndServe()
}

func (s *Server) StartDaemonLoop() error {
	configs, err := s.FileManager.LoadConfigs()
	if err != nil {
		return err
	}

	interval := time.Duration(configs.CheckInterval) * time.Minute

	s.mu.Lock()
	var oldDone <-chan struct{}
	if s.currentLoopControl != nil {
		s.currentLoopControl.Cancel()
		oldDone = s.currentLoopControl.Done
	}
	s.mu.Unlock()

	// Wait for old goroutine to finish before starting new one to avoid race on status
	if oldDone != nil {
		select {
		case <-oldDone:
		case <-time.After(5 * time.Second):
		}
	}

	s.mu.Lock()
	loopControl := s.StartLoopFunc(daemon.StartLoopPayload{
		FileManager: s.FileManager,
		Interval:    interval,
		State:       s.State,
		JobQueue:    s.JobQueue,
		Backend:     s.Torrents,
		Librarian:   s.Librarian,
	})
	s.currentLoopControl = loopControl
	s.mu.Unlock()

	return nil
}

func (s *Server) StopDaemonLoop() {
	s.mu.Lock()
	hadLoop := s.currentLoopControl != nil
	var oldDone <-chan struct{}
	if s.currentLoopControl != nil {
		s.currentLoopControl.Cancel()
		oldDone = s.currentLoopControl.Done
		s.currentLoopControl = nil
	}
	s.mu.Unlock()

	// Wait for the goroutine to finish so status is correctly set to stopped
	if oldDone != nil {
		select {
		case <-oldDone:
		case <-time.After(5 * time.Second):
		}
	}

	if hadLoop && s.State.GetStatus() != daemon.StatusStopped {
		s.State.SetStatus(daemon.StatusStopped)
	}
}

func (s *Server) handleWebSocket() http.HandlerFunc {
	return func(w http.ResponseWriter, r *http.Request) {
		s.WSManager.HandleWebSocket(w, r)
	}
}

func (s *Server) Stop(ctx context.Context) error {
	logger.Logger.Info().Msg("Stopping API server")

	// Close WebSocket connections
	if s.WSManager != nil {
		s.WSManager.Close()
	}

	return s.Shutdown(ctx)
}